Drop (implied) dependency on browsing  and webpage
The `browsing` and `webpage` apps are part of the apis-core Django project.
Given that apis-infos is now a separate app, we have no way to access
funcionality from them. We could depend on apis-core, but the functionality of
the apps that we need is simple and depending on a whole full blown project
just for a couple of views does not stand for it. Therefore we include the
relevant parts of `browsing` and `webpage` in this app (mostly some views
and templates).